devices: allow sending on device-add notifications for major/minor-less devices
This is usefull for getting device plug/unplug notifications for
things like some USB adapters which do not have major/minor
numbers.

Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>

+++ init/devices.c
@@ -518,7 +518,8 @@ static void handle_device(const char *action, const char *devpath,
     int i;
 
     if(!strcmp(action, "add")) {
-        make_device(devpath, path, block, major, minor);
+        if (major >= 0 && minor >= 0)
+            make_device(devpath, path, block, major, minor);
         if (links) {
             for (i = 0; links[i]; i++)
                 make_link(devpath, links[i]);
@@ -530,7 +531,8 @@ static void handle_device(const char *action, const char *devpath,
             for (i = 0; links[i]; i++)
                 remove_link(devpath, links[i]);
         }
-        unlink(devpath);
+        if (major >= 0 && minor >= 0)
+            unlink(devpath);
     }
 
     if (links) {
@@ -553,10 +555,6 @@ static void handle_platform_device_event(struct uevent *uevent)
 static const char *parse_device_name(struct uevent *uevent, unsigned int len)
 {
     const char *name;
-
-    /* if it's not a /dev device, nothing else to do */
-    if((uevent->major < 0) || (uevent->minor < 0))
-        return NULL;
 
     /* do we have a name? */
     name = strrchr(uevent->path, '/');
